+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changes
 
 - New: Add option to avoid overlapping of the circle / square symbols (in PropSymbol, PropSymbolChoro and PropSymbolTypo).
 
+- Update NUTS datasets to 2020 version.
+
 
 0.10.1 (2022-10-13)
 -------------------